functions/ml_event_handler/requirements.txt
google-cloud-logging
google-cloud-aiplatform>=1.38.0